backend/requirements.txt
Flask==2.2.5
boto3==1.26.145
gunicorn==20.1.0
flask_cors
python-dotenv